Actually clock_gettime is not libc
